Add registry functionality moved to a separate module.
